Quick and simple front-end file fix and adding language buttons logic
