+++ src/main/java/studos/gui/LoginWindow.java
@@ -6,18 +6,10 @@ package studos.gui;
 
 import java.io.IOException;
 import javafx.application.Application;
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.CheckBox;
-import javafx.scene.control.PasswordField;
-import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
-import javafx.scene.input.MouseEvent;
-import javafx.scene.text.Text;
 import javafx.stage.Stage;
 import javafx.stage.StageStyle;
 import studos.logic.AppLanguage;
@@ -37,15 +29,7 @@ public class LoginWindow extends Application {
     public static void main(final String[] args) {
         launch(args);
     }
-    /**
-     * Variable that store 'X' from login window position.
-     */
 
-    private double xOffset = 0;
-    /**
-     * Variable that store 'Y' from login window position.
-     */
-    private double yOffset = 0;
     /**
      * Method that runs front-end of login window.
      * @throws IOException IOexception
@@ -60,8 +44,7 @@ public class LoginWindow extends Application {
         UserConfigReader.readUserData();
         AppLanguage.languageInitialization();
         /*
-         * Setting window properties.
-         * Title, setting resizable to false
+         * Setting window properties. Title, setting resizable to false
          * and removing context menu from window.
         */
         primaryStage.setTitle("Studos - Okno logowania aplikacji");
@@ -70,7 +53,6 @@ public class LoginWindow extends Application {
         primaryStage.getIcons()
             .add(new Image(getClass()
             .getResourceAsStream("/loginWindow/logoIcon.png")));
-
         // Loading controlls from .fxml file and setting size of window.
         final FXMLLoader loader =
              new FXMLLoader(GuiStarter.class
@@ -79,107 +61,10 @@ public class LoginWindow extends Application {
         final int windowHeight = 494;
         final int windowWidth = 1110;
 
-        // Creating variables from content in front-end controlls.
-        final Button loginButton =
-                 (Button) loader.getNamespace().get("loginButton");
-        final Button minimalizeButton =
-                 (Button) loader.getNamespace().get("minimalizeButton");
-        final Button closeButton =
-                 (Button) loader.getNamespace().get("closeButton");
-        final TextField usernameText =
-                  (TextField) loader.getNamespace()
-                  .get("usernameTextField");
-        final PasswordField passwordText =
-                  (PasswordField) loader.getNamespace()
-                  .get("passwordTextField");
-        final Text loginMessageText =
-                  (Text) loader.getNamespace()
-                  .get("loginMessageText");
-        final CheckBox rememberCheckBox =
-                    (CheckBox) loader.getNamespace().get("rememberMeCheckbox");
-        //Todo - to backend file
-        if (UserConfigReader.ifDataIsReady()) {
-            try {
-                rememberCheckBox.setSelected(true);
-                usernameText.setText(UserConfigReader.getUsername());
-                passwordText.setText(UserConfigReader.getPassword());
-            } catch (final Exception e) {
-              }
-        }
-        /*
-         * Window draggable method.
-         * This action will allow user to drag window.
-        */
-        root.setOnMousePressed(new EventHandler<MouseEvent>() {
-            @Override
-            public void handle(final MouseEvent event) {
-                xOffset = event.getSceneX();
-                yOffset = event.getSceneY();
-                root.requestFocus();
-            }
-        });
-        root.setOnMouseDragged(new EventHandler<MouseEvent>() {
-            @Override
-            public void handle(final MouseEvent event) {
-                primaryStage.setX(event.getScreenX() - xOffset);
-                primaryStage.setY(event.getScreenY() - yOffset);
-            }
-        });
-        /*
-         * Window close button.
-         * This action will close app window.
-        */
-        closeButton.setOnAction(new EventHandler<ActionEvent>() {
-            @Override
-            public void handle(final ActionEvent e) {
-                primaryStage.close();
-            }
-        });
-        /*
-         * Window minimalize button.
-         * This action will minimalize app window.
-        */
-        minimalizeButton.setOnAction(new EventHandler<ActionEvent>() {
-            @Override
-            public void handle(final ActionEvent e) {
-                primaryStage.setIconified(true);
-            }
-        });
-        /*
-         * Button login execute.
-         * This action will run back-end login method.
-        */
-        loginButton.setOnAction(new EventHandler<ActionEvent>() {
-            @Override
-            public void handle(final ActionEvent e) {
-                final String username = usernameText.getText();
-                final String password = passwordText.getText();
-                /*
-                 * Function below is checking user credentials.
-                 * (login function)
-                 */
-                final LoginLogic loginLogic = new LoginLogic();
-                if (loginLogic.ifPlaceIsEmpty(username, password)) {
-                    loginMessageText
-                    .setText("Dane logowania nie zostały uzupełnione.");
-                } else {
-                    if (loginLogic.ifInitialIsRight(username, password)) {
-                        loginLogic.secondWindow(username, password);
-                        if (rememberCheckBox.isSelected()) {
-                            UserConfigReader.saveUserData(usernameText
-                                 .getText(), passwordText.getText(), "");
-                        } else {
-                            UserConfigReader.deleteUserData();
-                        }
-                        primaryStage.close();
-                    } else {
-                        loginMessageText
-                        .setText("Nieprawidłowe dane logowania.");
-                    }
-                }
-                loginButton.requestFocus();
-            }
-        });
+        // Temporary method that stores controlls
+        // and handle controlls actions
+        LoginLogic.controlls(loader, root, primaryStage);
+
         /*
          * Setting scene with settings declared above
          * and loading .css loginWindow file.
