+++ src/main/java/studos/gui/view/loginWindow.fxml
@@ -4,6 +4,8 @@
 <?import javafx.scene.control.CheckBox?>
 <?import javafx.scene.control.PasswordField?>
 <?import javafx.scene.control.TextField?>
+<?import javafx.scene.control.ToggleButton?>
+<?import javafx.scene.control.ToggleGroup?>
 <?import javafx.scene.image.Image?>
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
@@ -52,10 +54,14 @@
             </ImageView>
             <Button fx:id="closeButton" focusTraversable="false" layoutX="1065.0" layoutY="-1.0" mnemonicParsing="false" prefHeight="45.0" prefWidth="45.0" />
             <Button fx:id="minimalizeButton" focusTraversable="false" layoutX="1020.0" layoutY="-1.0" mnemonicParsing="false" prefHeight="45.0" prefWidth="45.0" />
-            <Button fx:id="polishLanguageButton" layoutX="913.0" layoutY="421.0" maxHeight="17.0" maxWidth="33.0" minHeight="17.0" minWidth="33.0" mnemonicParsing="false" prefHeight="17.0" prefWidth="33.0" />
-            <Button fx:id="englishLanguageButton" layoutX="990.0" layoutY="421.0" maxHeight="17.0" maxWidth="33.0" minHeight="17.0" minWidth="33.0" mnemonicParsing="false" prefHeight="17.0" prefWidth="33.0" />
             <Text fx:id="loginMessageText" layoutX="855.0" layoutY="305.0" strokeType="OUTSIDE" strokeWidth="0.0" textAlignment="CENTER" wrappingWidth="234.0" />
             <Button fx:id="settingsButton" focusTraversable="false" layoutX="977.0" layoutY="-1.0" mnemonicParsing="false" prefHeight="45.0" prefWidth="45.0" />
+            <ToggleButton fx:id="polishLanguageButton" layoutX="913.0" layoutY="421.0" maxHeight="17.0" maxWidth="33.0" minHeight="17.0" minWidth="33.0" mnemonicParsing="false" prefHeight="17.0" prefWidth="33.0">
+               <toggleGroup>
+                  <ToggleGroup fx:id="ToggleLanguages" />
+               </toggleGroup>
+            </ToggleButton>
+            <ToggleButton fx:id="englishLanguageButton" layoutX="990.0" layoutY="421.0" maxHeight="17.0" maxWidth="33.0" minHeight="17.0" minWidth="33.0" mnemonicParsing="false" prefHeight="17.0" prefWidth="33.0" toggleGroup="$ToggleLanguages" />
       </children>
     </AnchorPane>
   </children>
