+++ src/main/java/studos/logic/UserConfigReader.java
@@ -163,7 +163,7 @@ private UserConfigReader() {
 
       // language
       final Element language = document.createElement("language");
-      language.appendChild(document.createTextNode("pl"));
+      language.appendChild(document.createTextNode(AppLanguage.getLanguage()));
       root.appendChild(language);
 
       // create the xml file
